Breaking up functions and updating README.md
