+++ src/go_melt/core/computeFunctions.py
@@ -1,6 +1,3 @@
-import copy
-import json
-import os
 from functools import partial
 
 import dill
@@ -11,8 +8,10 @@ from jax.experimental import sparse
 from jax.numpy import multiply
 from pyevtk.hl import gridToVTK
 import sys
-
-# TFSP: Temporary fix for single precision
+from go_melt.utils.gaussian_quadrature import (
+    computeQuad2dFemShapeFunctions_jax,
+    computeQuad3dFemShapeFunctions_jax,
+)
 
 
 def calcNumNodes(elements):
@@ -64,40 +63,6 @@ def createMesh3D(x, y, z):
     return [nx, ny, nz], [nconn_x, nconn_y, nconn_z]
 
 
-class obj:
-    """
-    A simple wrapper class to convert a dictionary into an object
-    with attributes accessible via dot notation.
-
-    Attributes are dynamically created from the keys and values
-    of the input dictionary.
-
-    Example:
-        data = {'a': 1, 'b': 2}
-        o = obj(data)
-        print(o.a)  # Outputs: 1
-    """
-
-    def __init__(self, dict1):
-        self.__dict__.update(dict1)
-
-
-def dict2obj(dict1):
-    """
-    Convert a dictionary into an object with attribute-style access.
-
-    This function serializes the dictionary to JSON and then deserializes
-    it using a custom object hook to create an instance of the `obj` class.
-
-    Parameters:
-    dict1 (dict): The dictionary to convert.
-
-    Returns:
-    obj: An object with attributes corresponding to the dictionary keys.
-    """
-    return json.loads(json.dumps(dict1), object_hook=obj)
-
-
 def save_object(obj, filename):
     """
     Save a Python object to a file using the dill serialization library.
@@ -110,338 +75,6 @@ def save_object(obj, filename):
     """
     with open(filename, "wb") as outp:
         dill.dump(obj, outp, dill.HIGHEST_PROTOCOL)
-
-
-def SetupLevels(solver_input, properties):
-    """
-    Set up multilevel mesh structures and initialize physical fields.
-
-    This function constructs four hierarchical mesh levels (Level0-Level3),
-    computes geometric and physical properties, and identifies overlapping
-    regions between levels for multiscale simulations. Level0 is fine-scale
-    top layer. Level 1 is part-scale coarse mesh. Level 2 is meso-scale.
-    Level 3 is fine-scale near the melt pool.
-
-    Parameters:
-    solver_input (dict): Dictionary containing configuration for Level1-Level3.
-    properties (dict): Dictionary of physical and simulation parameters.
-
-    Returns:
-    list: A list of dictionaries representing Level0 to Level3 configurations.
-    """
-
-    # Parse solver input into Level1–Level3 objects
-    Level1 = dict2obj(solver_input.get("Level1", {}))
-    Level2 = dict2obj(solver_input.get("Level2", {}))
-    Level3 = dict2obj(solver_input.get("Level3", {}))
-
-    # Common setup for all three levels
-    for level in [Level1, Level2, Level3]:
-        level.length, level.h = calc_length_h(level)
-        level.nodes = calcNumNodes(level.elements)
-        level.ne = level.elements[0] * level.elements[1] * level.elements[2]
-        level.nn = level.nodes[0] * level.nodes[1] * level.nodes[2]
-        level.BC = getBCindices(level)
-        level.node_coords, level.connect = createMesh3D(
-            (level.bounds.x[0], level.bounds.x[1], level.nodes[0]),
-            (level.bounds.y[0], level.bounds.y[1], level.nodes[1]),
-            (level.bounds.z[0], level.bounds.z[1], level.nodes[2]),
-        )
-        level.T = properties["T_amb"] * jnp.ones(level.nn)
-        level.T0 = properties["T_amb"] * jnp.ones(level.nn)
-        level.S1 = jnp.zeros(level.nn)
-        level.S2 = jnp.zeros(level.nn, dtype=bool)
-        level.k = properties["k_powder"] * jnp.ones(level.nn)
-        level.rhocp = (
-            properties["cp_solid_coeff_a0"] * properties["rho"] * jnp.ones(level.nn)
-        )
-
-    # Special storage for Level1
-    Level1.S1_storage = jnp.zeros(
-        [int(round(Level1.h[2] / properties["layer_height"])), Level1.nn]
-    )
-
-    # Additional setup for Level2 and Level3
-    for level in [Level2, Level3]:
-        (
-            level.bounds.ix,
-            level.bounds.iy,
-            level.bounds.iz,
-        ) = find_max_const(Level1, level)
-        level.init_node_coors = copy.deepcopy(level.node_coords)
-        level.Tprime = jnp.zeros(level.nn)
-        level.Tprime0 = copy.deepcopy(level.Tprime)
-
-    # Adjust Level1 z-coordinates to align with Level2
-    Level1.orig_node_coords = copy.deepcopy(Level1.node_coords)
-    tmp_coords = copy.deepcopy(Level1.orig_node_coords)
-    while True:
-        if jnp.isclose(tmp_coords[2] - Level2.node_coords[-1][-1], 0, atol=1e-4).any():
-            break
-        tmp_coords[2] += properties["layer_height"]
-    Level1.node_coords = copy.deepcopy(tmp_coords)
-
-    # Overlap regions between levels
-    def get_overlap(level_fine, level_coarse):
-        nodes = [
-            getCoarseNodesInFineRegion(
-                level_fine.node_coords[i], level_coarse.node_coords[i]
-            )
-            for i in range(3)
-        ]
-        coors = [
-            jnp.array([level_coarse.node_coords[i][j] for j in nodes[i]])
-            for i in range(3)
-        ]
-        return nodes, coors
-
-    Level2.orig_overlap_nodes, Level2.orig_overlap_coors = get_overlap(Level2, Level1)
-    Level3.orig_overlap_nodes, Level3.orig_overlap_coors = get_overlap(Level3, Level2)
-
-    Level2.overlapNodes = copy.deepcopy(Level2.orig_overlap_nodes)
-    Level2.overlapCoords = copy.deepcopy(Level2.orig_overlap_coors)
-    Level3.overlapNodes = copy.deepcopy(Level3.orig_overlap_nodes)
-    Level3.overlapCoords = copy.deepcopy(Level3.orig_overlap_coors)
-
-    # Create Level0 for high-resolution state tracking
-    Level0 = obj({})
-    Level0.elements = [
-        round(Level1.length[0] / Level3.h[0]),
-        round(Level1.length[1] / Level3.h[1]),
-        round(Level2.length[2] / Level3.h[2]),
-    ]
-    Level0.nodes = calcNumNodes(Level0.elements)
-    Level0.ne = Level0.elements[0] * Level0.elements[1] * Level0.elements[2]
-    Level0.nn = Level0.nodes[0] * Level0.nodes[1] * Level0.nodes[2]
-    Level0.node_coords, Level0.connect = createMesh3D(
-        (Level1.bounds.x[0], Level1.bounds.x[1], Level0.nodes[0]),
-        (Level1.bounds.y[0], Level1.bounds.y[1], Level0.nodes[1]),
-        (Level2.bounds.z[0], Level2.bounds.z[1], Level0.nodes[2]),
-    )
-    Level0.orig_node_coords = copy.deepcopy(Level0.node_coords)
-
-    Level0.orig_overlap_nodes, Level0.orig_overlap_coors = get_overlap(Level3, Level0)
-
-    Level0.overlapNodes = copy.deepcopy(Level0.orig_overlap_nodes)
-    Level0.overlapCoords = copy.deepcopy(Level0.orig_overlap_coors)
-
-    Level0.orig_overlap_nodes_L2 = [
-        getCoarseNodesInLargeFineRegion(Level2.node_coords[i], Level0.node_coords[i])
-        for i in range(3)
-    ]
-    Level0.orig_overlap_coors_L2 = [
-        jnp.array([Level0.node_coords[i][j] for j in Level0.orig_overlap_nodes_L2[i]])
-        for i in range(3)
-    ]
-    Level0.overlapNodes_L2 = copy.deepcopy(Level0.orig_overlap_nodes_L2)
-    Level0.overlapCoords_L2 = copy.deepcopy(Level0.orig_overlap_coors_L2)
-
-    Level0.S1 = jnp.zeros(Level0.nn)
-    Level0.S2 = jnp.zeros(Level0.nn, dtype=bool)
-    Level0.idx = getOverlapRegion(Level0.overlapNodes, Level0.nodes[0], Level0.nodes[1])
-    Level0.idx_L2 = getOverlapRegion(
-        Level0.overlapNodes_L2, Level0.nodes[0], Level0.nodes[1]
-    )
-
-    Level0.layer_idx_delta = int(round(properties["layer_height"] / Level3.h[2]))
-
-    # Convert all numeric fields to jnp arrays
-    for level in [Level0, Level1, Level2, Level3]:
-        for attr in ["elements", "h", "length", "nodes"]:
-            if hasattr(level, attr):
-                setattr(level, attr, [jnp.array(v) for v in getattr(level, attr)])
-        for attr in ["ne", "nn", "layer_idx_delta"]:
-            if hasattr(level, attr):
-                setattr(level, attr, jnp.array(getattr(level, attr)))
-
-    # Return all levels as dictionaries
-    Levels = [
-        structure_to_dict(Level0),
-        structure_to_dict(Level1),
-        structure_to_dict(Level2),
-        structure_to_dict(Level3),
-    ]
-    return Levels
-
-
-def SetupProperties(prop_obj):
-    """
-    Initialize and return a dictionary of material and simulation properties.
-
-    This function reads user-defined or default values from the input
-    dictionary and assigns them to a structured object for use in
-    simulations. It also computes derived coefficients used in
-    evaporation and heat transfer models.
-
-    Parameters:
-    prop_obj (dict): Dictionary containing material and simulation parameters.
-
-    Returns:
-    dict: A dictionary of structured properties with default fallbacks.
-    """
-    properties = dict2obj(prop_obj)
-
-    # --- Thermal conductivity (W/m·K) ---
-    # Powder: constant
-    properties.k_powder = prop_obj.get("thermal_conductivity_powder", 0.4)
-    # Bulk: linear temperature dependence (a1 * T + a0)
-    properties.k_bulk_coeff_a0 = prop_obj.get("thermal_conductivity_bulk_a0", 4.23)
-    properties.k_bulk_coeff_a1 = prop_obj.get("thermal_conductivity_bulk_a1", 0.016)
-    # Fluid: constant
-    properties.k_fluid_coeff_a0 = prop_obj.get("thermal_conductivity_fluid_a0", 29.0)
-
-    # --- Heat capacity (J/kg·K) ---
-    # Solid: linear temperature dependence
-    properties.cp_solid_coeff_a0 = prop_obj.get("heat_capacity_solid_a0", 383.1)
-    properties.cp_solid_coeff_a1 = prop_obj.get("heat_capacity_solid_a1", 0.174)
-    # Mushy and fluid phases: constant
-    properties.cp_mushy = prop_obj.get("heat_capacity_mushy", 3235.0)
-    properties.cp_fluid = prop_obj.get("heat_capacity_fluid", 769.0)
-
-    # --- Density (kg/mm³) ---
-    properties.rho = prop_obj.get("density", 8.0e-6)
-
-    # --- Laser parameters ---
-    properties.laser_radius = prop_obj.get("laser_radius", 0.110)  # mm
-    properties.laser_depth = prop_obj.get("laser_depth", 0.05)  # mm
-    properties.laser_power = prop_obj.get("laser_power", 300.0)  # W
-    properties.laser_eta = prop_obj.get("laser_absorptivity", 0.25)  # unitless
-    properties.laser_center = prop_obj.get("laser_center", [])  # mm
-
-    # --- Temperature thresholds (K) ---
-    properties.T_amb = prop_obj.get("T_amb", 353.15)
-    properties.T_solidus = prop_obj.get("T_solidus", 1554.0)
-    properties.T_liquidus = prop_obj.get("T_liquidus", 1625.0)
-    properties.T_boiling = prop_obj.get("T_boiling", 3038.0)
-
-    # --- Heat transfer and radiation ---
-    properties.h_conv = prop_obj.get("h_conv", 1.473e-5)  # W/mm²·K
-    properties.h_conv *= 1e6 # Quick fix: later assumes in m for surface BC calc
-    properties.vareps = prop_obj.get("emissivity", 0.600)  # unitless
-    properties.evc = prop_obj.get("evaporation_coefficient", 0.82)  # unitless
-
-    # --- Physical constants ---
-    properties.kb = prop_obj.get("boltzmann_constant", 1.38e-23)  # J/K
-    properties.mA = prop_obj.get("atomic_mass", 7.9485017e-26)  # kg
-    properties.Lev = prop_obj.get("latent_heat_evap", 4.22e6)  # J/kg
-    properties.molar_mass = prop_obj.get("molar_mass", 58.69) * 1e-3  # g/mol → kg/mol
-
-    # --- Layer height (mm) ---
-    properties.layer_height = prop_obj.get("layer_height", 0.04)
-
-    # --- Universal constants ---
-    properties.sigma_sb = 5.67e-8  # Stefan–Boltzmann constant (W/m²·K⁴)
-    properties.gas_const = 8.314  # Molar gas constant (J/mol·K)
-    properties.atmospheric_pressure = 101325  # Pa
-
-    # --- Derived evaporation model coefficients ---
-    # CM_coeff: Heat loss temperature factor (K·s²/m²)
-    properties.CM_coeff = properties.molar_mass / (2.0 * jnp.pi * properties.gas_const)
-    # CT_coeff: Recoil pressure temperature factor (K)
-    properties.CT_coeff = properties.Lev * properties.molar_mass / properties.gas_const
-    # CP_coeff: Recoil pressure factor (Pa)
-    properties.CP_coeff = 0.54 * properties.atmospheric_pressure
-
-    return structure_to_dict(properties)
-
-
-def SetupNonmesh(nonmesh_input):
-    """
-    Initialize and return a dictionary of non-mesh simulation parameters.
-
-    Parameters:
-    nonmesh_input (dict): Dictionary containing time-stepping, output,
-                          and laser path configuration.
-
-    Returns:
-    dict: A dictionary of structured non-mesh parameters.
-    """
-    Nonmesh = dict2obj(nonmesh_input)
-
-    # Timestep for finest mesh (s)
-    Nonmesh.timestep_L3 = nonmesh_input.get("timestep_L3", 1e-5)
-
-    # Subcycle numbers (unitless)
-    Nonmesh.subcycle_num_L2 = nonmesh_input.get("subcycle_num_L2", 1)
-    Nonmesh.subcycle_num_L3 = nonmesh_input.get("subcycle_num_L3", 1)
-
-    # Dwell time duration (s)
-    Nonmesh.dwell_time = nonmesh_input.get("dwell_time", 0.1)
-
-    # Coarse-domain recording frequency (steps)
-    Nonmesh.Level1_record_step = nonmesh_input.get("Level1_record_step", 1)
-
-    # Output directory
-    Nonmesh.save_path = nonmesh_input.get("save_path", "results/")
-
-    # Output flags
-    Nonmesh.output_files = nonmesh_input.get("output_files", 1)  # Save VTK/VTR
-
-    # Toolpath file location
-    Nonmesh.toolpath = nonmesh_input.get("toolpath", "laserPath.txt")
-
-    # Wait time before increasing timestep in dwell (steps)
-    Nonmesh.wait_time = nonmesh_input.get("wait_time", 500.0)
-
-    # Layer control
-    Nonmesh.layer_num = nonmesh_input.get("layer_num", 0)
-    Nonmesh.restart_layer_num = nonmesh_input.get("restart_layer_num", 10000)
-
-    # Output verbosity
-    Nonmesh.info_T = nonmesh_input.get("info_T", 0)
-
-    # Laser velocity (mm/s)
-    Nonmesh.laser_velocity = nonmesh_input.get("laser_velocity", 500)
-
-    # Wait time after each track (s)
-    Nonmesh.wait_track = nonmesh_input.get("wait_track", 0.0)
-
-    # Recording frequency (steps)
-    Nonmesh.record_step = nonmesh_input.get(
-        "record_step", Nonmesh.subcycle_num_L2 * Nonmesh.subcycle_num_L3
-    )
-
-    # G-code file path
-    Nonmesh.gcode = nonmesh_input.get(
-        "gcode", "./examples/gcodefiles/defaultName.gcode"
-    )
-
-    # Dwell time multiplier (unitless)
-    Nonmesh.dwell_time_multiplier = nonmesh_input.get("dwell_time_multiplier", 1)
-
-    # Use TXT format for toolpath (flag)
-    Nonmesh.use_txt = nonmesh_input.get("use_txt", 0)
-
-    # Create output directory if it doesn't exist
-    if not os.path.exists(Nonmesh.save_path):
-        os.makedirs(Nonmesh.save_path)
-
-    return structure_to_dict(Nonmesh)
-
-
-def structure_to_dict(struct):
-    """
-    Recursively convert a nested object with attributes into a dictionary.
-
-    This function handles nested objects by checking if each attribute
-    has a __dict__ (i.e., is an object) and recursively converting it,
-    unless the object has a 'tolist' method (e.g., NumPy or JAX arrays),
-    in which case it is left as-is.
-
-    Parameters:
-    struct (object): An object with attributes to convert.
-
-    Returns:
-    dict: A dictionary representation of the object's attributes.
-    """
-    return {
-        k: (
-            v
-            if (not hasattr(v, "__dict__") or hasattr(v, "tolist"))
-            else structure_to_dict(v)
-        )
-        for k, v in struct.__dict__.items()
-    }
 
 
 def getStaticNodesAndElements(L):
@@ -689,172 +322,6 @@ def convert2XYZ(i, ne_x, ne_y, nn_x, nn_y):
     return ix, iy, iz, idx
 
 
-@jax.jit
-def computeQuad3dFemShapeFunctions_jax(coords):
-    """
-    Compute shape functions, their derivatives, and quadrature weights
-    for an 8-node hexahedral element using 3D Gaussian quadrature.
-
-    This function evaluates the shape function matrix (N), its spatial
-    derivatives (dNdx), and the quadrature weights (wq) at 8 integration
-    points in the reference element.
-
-    Parameters:
-    coords (array): Nodal coordinates of the hexahedral element, shape (8, 3).
-
-    Returns:
-    N (array): Shape function values at each quadrature point, shape (8, 8).
-    dNdx (array): Derivatives of shape functions w.r.t. global coordinates,
-                  shape (8, 8, 3).
-    wq (array): Quadrature weights for each integration point, shape (8, 1).
-    """
-    ngp = 8  # Number of Gauss points
-    ndim = 3  # Number of spatial dimensions
-
-    # Isoparametric coordinates for 8-node hexahedral element
-    ksi_i = jnp.array([-1, 1, 1, -1, -1, 1, 1, -1])
-    eta_i = jnp.array([-1, -1, 1, 1, -1, -1, 1, 1])
-    zeta_i = jnp.array([-1, -1, -1, -1, 1, 1, 1, 1])
-
-    # Gauss points in reference space (±1/√3)
-    ksi_q = (1 / jnp.sqrt(3)) * ksi_i
-    eta_q = (1 / jnp.sqrt(3)) * eta_i
-    zeta_q = (1 / jnp.sqrt(3)) * zeta_i
-
-    # Uniform quadrature weights for 2-point Gauss rule
-    tmp_wq = jnp.ones(ngp)
-
-    # Evaluate shape functions at Gauss points
-    _ksi = 1 + ksi_q[:, None] @ ksi_i[None, :]
-    _eta = 1 + eta_q[:, None] @ eta_i[None, :]
-    _zeta = 1 + zeta_q[:, None] @ zeta_i[None, :]
-    N = (1 / 8) * _ksi * _eta * _zeta
-    # Derivatives w.r.t. reference coordinates
-    dNdksi = (1 / 8) * ksi_i[None, :] * _eta * _zeta
-    dNdeta = (1 / 8) * eta_i[None, :] * _ksi * _zeta
-    dNdzeta = (1 / 8) * zeta_i[None, :] * _ksi * _eta
-
-    # Compute Jacobian components (diagonal for structured mesh)
-    dxdksi = jnp.matmul(dNdksi, coords[:, 0])
-    dydeta = jnp.matmul(dNdeta, coords[:, 1])
-    dzdzeta = jnp.matmul(dNdzeta, coords[:, 2])
-
-    # Manually construct Jacobian and its inverse
-    J = jnp.array(
-        [
-            [dxdksi[0], 0.0, 0.0],
-            [0.0, dydeta[0], 0.0],
-            [0.0, 0.0, dzdzeta[0]],
-        ]
-    )
-    Jinv = jnp.array(
-        [
-            [1.0 / dxdksi[0], 0.0, 0.0],
-            [0.0, 1.0 / dydeta[0], 0.0],
-            [0.0, 0.0, 1.0 / dzdzeta[0]],
-        ]
-    )
-
-    # Allocate arrays for shape function derivatives and weights
-    dNdx = jnp.zeros((ngp, ngp, ndim))
-    wq = jnp.zeros((ngp, 1))
-
-    # Loop over Gauss points to compute global derivatives and weights
-    for q in range(ngp):
-        dN_dxi = jnp.concatenate(
-            [
-                dNdksi[q, :, None],
-                dNdeta[q, :, None],
-                dNdzeta[q, :, None],
-            ],
-            axis=1,
-        )
-        dNdx = dNdx.at[q, :, :].set(dN_dxi @ Jinv)
-        wq = wq.at[q].set(jnp.linalg.det(J) * tmp_wq[q])
-
-    return jnp.array(N), jnp.array(dNdx), jnp.array(wq)
-
-
-@jax.jit
-def computeQuad2dFemShapeFunctions_jax(coords):
-    """
-    Compute shape functions, their derivatives, and quadrature weights
-    for a 4-node quadrilateral element using 2D Gaussian quadrature.
-
-    This function evaluates the shape function matrix (N), its spatial
-    derivatives (dNdx), and the quadrature weights (wq) at 4 integration
-    points in the reference element.
-
-    Parameters:
-    coords (array): Nodal coordinates of the quadrilateral element, shape (8, 2). Only
-                    the last 4 coordinates (top surface of element) are used.
-
-    Returns:
-    N (array): Shape function values at each quadrature point, shape (4, 4).
-    dNdx (array): Derivatives of shape functions w.r.t. global coordinates,
-                  shape (4, 4, 2).
-    wq (array): Quadrature weights for each integration point, shape (4, 1).
-    """
-    ngp = 4  # Number of Gauss points
-    ndim = 2  # Number of spatial dimensions
-
-    # Isoparametric coordinates for 4-node quadrilateral element
-    ksi_i = jnp.array([-1, 1, 1, -1])
-    eta_i = jnp.array([-1, -1, 1, 1])
-
-    # Gauss points in reference space (±1/√3)
-    ksi_q = (1 / jnp.sqrt(3)) * ksi_i
-    eta_q = (1 / jnp.sqrt(3)) * eta_i
-
-    # Uniform quadrature weights for 2-point Gauss rule
-    tmp_wq = jnp.ones(ngp)
-
-    # Evaluate shape functions at Gauss points
-    _ksi = 1 + ksi_q[:, None] @ ksi_i[None, :]
-    _eta = 1 + eta_q[:, None] @ eta_i[None, :]
-    N = (1 / 4) * _ksi * _eta
-
-    # Derivatives w.r.t. reference coordinates
-    dNdksi = (1 / 4) * ksi_i[None, :] * _eta
-    dNdeta = (1 / 4) * eta_i[None, :] * _ksi
-
-    # Compute Jacobian components (diagonal for structured mesh)
-    dxdksi = jnp.matmul(dNdksi, coords[4:, 0])
-    dydeta = jnp.matmul(dNdeta, coords[4:, 1])
-
-    # Manually construct Jacobian and its inverse
-    J = jnp.array(
-        [
-            [dxdksi[0], 0.0],
-            [0.0, dydeta[0]],
-        ]
-    )
-    Jinv = jnp.array(
-        [
-            [1.0 / dxdksi[0], 0.0],
-            [0.0, 1.0 / dydeta[0]],
-        ]
-    )
-
-    # Allocate arrays for shape function derivatives and weights
-    dNdx = jnp.zeros((ngp, ngp, ndim))
-    wq = jnp.zeros((ngp, 1))
-
-    # Loop over Gauss points to compute global derivatives and weights
-    for q in range(ngp):
-        dN_dxi = jnp.concatenate(
-            [
-                dNdksi[q, :, None],
-                dNdeta[q, :, None],
-            ],
-            axis=1,
-        )
-        dNdx = dNdx.at[q, :, :].set(dN_dxi @ Jinv)
-        wq = wq.at[q].set(jnp.linalg.det(J) * tmp_wq[q])
-
-    return jnp.array(N), jnp.array(dNdx), jnp.array(wq)
-
-
 def getCoarseNodesInFineRegion(xnf, xnc):
     """
     Identify coarse grid nodes that overlap with the fine grid region.
@@ -923,106 +390,6 @@ def getCoarseNodesInLargeFineRegion(xnc, xnf):
     overlap = jnp.arange(overlapMin, overlapMax, step).astype(int)
 
     return overlap
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
-def computeSources(Level, v, Shapes, ne_nn, properties, laserP):
-    """
-    Compute integrated source terms for all three levels using Level 3 mesh.
-
-    This function evaluates the laser-induced heat source at quadrature
-    points of the fine mesh (Level 3), integrates it using shape functions,
-    and projects the result to coarser levels using precomputed transfer
-    operators.
-
-    Parameters:
-    Level (dict): Mesh level containing connectivity and geometry for Level 3.
-    v (array): Current laser position.
-    Shapes (list): Shape transfer operators between Level 3 and Levels 1 & 2.
-                   Shapes[1][0]: L3 → L1 interpolation matrix.
-                   Shapes[2][0]: L3 → L2 interpolation matrix.
-                   Shapes[1][2], Shapes[2][2]: Projection matrices.
-    ne_nn (tuple): Element/node counts for each level.
-                   ne_nn[1]: Number of elements in Level 3.
-                   ne_nn[4]: Number of nodes in Level 3.
-    properties (dict): Material and laser properties.
-    laserP (float): Laser power.
-
-    Returns:
-    Fc (array): Integrated source term for Level 1.
-    Fm (array): Integrated source term for Level 2.
-    Ff (array): Integrated source term for Level 3.
-    """
-    # Get shape functions and quadrature weights for Level 3 elements
-    coords = getSampleCoords(Level)
-    Nf, _, wqf = computeQuad3dFemShapeFunctions_jax(coords)
-
-    def stepcomputeCoarseSource(ieltf):
-        # Get nodal indices for the fine element
-        ix, iy, iz, idx = convert2XYZ(
-            ieltf,
-            Level["elements"][0],
-            Level["elements"][1],
-            Level["nodes"][0],
-            Level["nodes"][1],
-        )
-        # Evaluate coordinates at quadrature points
-        x, y, z = getQuadratureCoords(Level, ix, iy, iz, Nf)
-        # Evaluate source function at quadrature points
-        Q = computeSourceFunction_jax(x, y, z, v, properties, laserP)
-        # Return raw source, integrated source, and node indices
-        return Q * wqf, Nf @ Q * wqf, idx
-
-    # Vectorized computation over all Level 3 elements
-    vstepcomputeCoarseSource = jax.vmap(stepcomputeCoarseSource)
-    _data, _data3, nodes3 = vstepcomputeCoarseSource(jnp.arange(ne_nn[1]))
-
-    # Project source terms to Levels 1 and 2 using transfer operators
-    _data1 = multiply(Shapes[1][0], _data).sum(axis=1)
-    _data2 = multiply(Shapes[2][0], _data).sum(axis=1)
-
-    Fc = Shapes[1][2] @ _data1.reshape(-1)
-    Fm = Shapes[2][2] @ _data2.reshape(-1)
-    Ff = bincount(nodes3.reshape(-1), _data3.reshape(-1), ne_nn[4])
-
-    return Fc, Fm, Ff
-
-
-@jax.jit
-def computeSourceFunction_jax(x, y, z, v, properties, P):
-    """
-    Compute a 3D Gaussian heat source term for laser-material interaction.
-
-    This function evaluates a separable Gaussian distribution in x, y, and z
-    directions centered at the laser position `v`, scaled by laser power and
-    material properties.
-
-    Parameters:
-    x, y, z (array): Coordinates of quadrature points.
-    v (array): Laser center position [vx, vy, vz].
-    properties (dict): Material and laser properties.
-                       Required keys:
-                       - "laser_eta": laser absorption efficiency.
-                       - "laser_radius": standard deviation in x/y.
-                       - "laser_depth": standard deviation in z.
-    P (float): Laser power.
-
-    Returns:
-    array: Evaluated source term at each quadrature point.
-    """
-    # Precompute constants
-    _pcoeff = 6 * jnp.sqrt(3) * P * properties["laser_eta"]
-    _rcoeff = 1 / (properties["laser_radius"] * jnp.sqrt(jnp.pi))
-    _dcoeff = 1 / (properties["laser_depth"] * jnp.sqrt(jnp.pi))
-    _rsq = properties["laser_radius"] ** 2
-    _dsq = properties["laser_depth"] ** 2
-
-    # Evaluate separable Gaussian in each direction
-    Qx = _rcoeff * jnp.exp(-3 * (x - v[0]) ** 2 / _rsq)
-    Qy = _rcoeff * jnp.exp(-3 * (y - v[1]) ** 2 / _rsq)
-    Qz = _dcoeff * jnp.exp(-3 * (z - v[2]) ** 2 / _dsq)
-
-    return _pcoeff * Qx * Qy * Qz
 
 
 @jax.jit
@@ -2301,102 +1668,6 @@ def computeConvRadBC(Level, LevelT0, ne, nn, properties, F):
     return F + NeumannBC
 
 
-@partial(jax.jit, static_argnames=["ne_nn", "tmp_ne_nn", "substrate"])
-def stepGOMELT(
-    Levels, ne_nn, tmp_ne_nn, Shapes, LInterp, v, properties, dt, laserP, substrate
-):
-    """
-    Perform a full explicit time step for the multilevel GOMELT simulation.
-
-    This function executes the following sequence:
-      1. Update material state and thermal properties.
-      2. Compute source terms (laser, convection, radiation) for all levels.
-      3. Compute volumetric correction terms from previous Tprime.
-      4. Predictor step: solve temperature fields for all levels.
-      5. Compute new Tprime fields and updated volumetric corrections.
-      6. Corrector step: solve temperature fields again using updated corrections.
-      7. Update Tprime and temperature fields for the next time step.
-
-    Parameters:
-    Levels (dict): Multilevel mesh and field data.
-    ne_nn (list): Total number of elements and nodes for each level.
-    tmp_ne_nn (list): Active element/node counts for Level 1 (based on layer).
-    Shapes (list): Shape function data for interpolation between levels.
-    LInterp (list): Interpolation matrices between levels.
-    v (array): Current laser position.
-    properties (dict): Material and simulation properties.
-    dt (float): Time step size.
-    laserP (float): Laser power.
-    substrate (array): Node indices defining the substrate region.
-
-    Returns:
-    tuple:
-        - Levels (dict): Updated multilevel data with new temperatures and Tprime fields.
-        - _resetmask (array): Boolean mask indicating newly activated melt regions.
-    """
-    # Store previous melt state for comparison
-    preS2 = Levels[3]["S2"]
-
-    # Update material state and thermal properties
-    Levels, Lk, Lrhocp = updateStateProperties(Levels, properties, substrate)
-
-    # Unpack levels for clarity
-    L1, L2, L3 = Levels[1], Levels[2], Levels[3]
-
-    # --- Compute source terms ---
-    Fc, Fm, Ff = computeSources(L3, v, Shapes, ne_nn, properties, laserP)
-    Fc = computeConvRadBC(L1, L1["T0"], tmp_ne_nn[0], ne_nn[2], properties, Fc)
-    Fm = computeConvRadBC(L2, L2["T0"], ne_nn[0], ne_nn[3], properties, Fm)
-    Ff = computeConvRadBC(L3, L3["T0"], ne_nn[1], ne_nn[4], properties, Ff)
-    F = [0, Fc, Fm, Ff]  # Source terms for Levels 1–3
-
-    # --- Predictor step ---
-    Vcu, Vmu = computeCoarseTprimeTerm_jax(Levels, Lk[3], Lk[2], Shapes)
-    L1T, L2T, L3T = computeSolutions(
-        Levels, ne_nn, tmp_ne_nn, F, Vcu, LInterp, Lk, Lrhocp, Vmu, dt, properties
-    )
-
-    # Enforce minimum temperature (TFSP)
-    L1T = jnp.maximum(properties["T_amb"], L1T)
-    L2T = jnp.maximum(properties["T_amb"], L2T)
-    L3T = jnp.maximum(properties["T_amb"], L3T)
-
-    # --- Compute new Tprime fields ---
-    L3Tp, L2Tp, L2T, L1T = getBothNewTprimes(
-        Levels, L3T, L2T, LInterp[1], L1T, LInterp[0]
-    )
-
-    # --- Update volumetric correction terms ---
-    Vcu, Vmu = computeCoarseTprimeMassTerm_jax(
-        Levels, L3Tp, L2Tp, Lrhocp[3], Lrhocp[2], dt, Shapes, Vcu, Vmu
-    )
-
-    # --- Corrector step ---
-    L1T, L2T, Levels[3]["T0"] = computeSolutions(
-        Levels, ne_nn, tmp_ne_nn, F, Vcu, LInterp, Lk, Lrhocp, Vmu, dt, properties
-    )
-
-    # Enforce minimum temperature again (TFSP)
-    L1T = jnp.maximum(properties["T_amb"], L1T)
-    L2T = jnp.maximum(properties["T_amb"], L2T)
-    Levels[3]["T0"] = jnp.maximum(properties["T_amb"], Levels[3]["T0"])
-
-    # --- Final Tprime update for next time step ---
-    Levels[3]["Tprime0"], Levels[2]["Tprime0"], Levels[2]["T0"], Levels[1]["T0"] = (
-        getBothNewTprimes(Levels, Levels[3]["T0"], L2T, LInterp[1], L1T, LInterp[0])
-    )
-
-    # --- Update global state arrays ---
-    Levels[0]["S1"] = Levels[0]["S1"].at[Levels[0]["idx"]].set(Levels[3]["S1"])
-    Levels[0]["S2"] = Levels[0]["S2"].at[:].set(False)
-    Levels[0]["S2"] = Levels[0]["S2"].at[Levels[0]["idx"]].set(Levels[3]["S2"])
-
-    # Identify newly activated melt regions
-    _resetmask = ((1 - 2 * preS2) * Levels[3]["S2"]) == 1
-
-    return Levels, _resetmask
-
-
 @jax.jit
 def moveEverything(v, vstart, Levels, move_v, LInterp, L1L2Eratio, L2L3Eratio, height):
     """
@@ -2665,72 +1936,6 @@ def stepGOMELTDwellTime(Levels, tmp_ne_nn, ne_nn, properties, dt, substrate):
 
 
 @partial(jax.jit, static_argnames=["ne_nn"])
-def computeLevelSource(Levels, ne_nn, laser_position, LevelShape, properties, laserP):
-    """
-    Compute the volumetric heat source for Level 1 or Level 2 using Level 3 mesh.
-
-    This function evaluates the laser heat source on the Level 3 mesh using
-    quadrature integration, then projects the result onto the target level
-    (Level 1 or Level 2) using the provided shape function matrices.
-
-    Parameters:
-    Levels (list of dict): Multilevel simulation state. Level 3 is used for integration.
-    ne_nn (tuple): Mesh metadata.
-        - ne_nn[1] (int): Number of elements in Level 3.
-    laser_position (array): Laser center positions for each time step.
-    LevelShape (tuple): FEM shape function data for the target level.
-        - LevelShape[0] (array): Shape function values at quadrature points.
-        - LevelShape[2] (array): Projection matrix to target level's global vector.
-    properties (dict): Material and process properties.
-    laserP (array): Laser power or intensity at each time step.
-
-    Returns:
-    array: Assembled global heat source vector for the target level.
-    """
-    # Get quadrature shape functions and weights for Level 3
-    coords = getSampleCoords(Levels[3])
-    Nf, _, wqf = computeQuad3dFemShapeFunctions_jax(coords)
-
-    def stepLaserPosition(ilaser):
-        def stepcomputeCoarseSource(ieltf):
-            # Convert element index to 3D indices and global node index
-            ix, iy, iz, idx = convert2XYZ(
-                ieltf,
-                Levels[3]["elements"][0],
-                Levels[3]["elements"][1],
-                Levels[3]["nodes"][0],
-                Levels[3]["nodes"][1],
-            )
-            # Get quadrature point coordinates for this element
-            x, y, z = getQuadratureCoords(Levels[3], ix, iy, iz, Nf)
-            w = wqf
-
-            # Evaluate laser source at quadrature points
-            Q = computeSourceFunction_jax(
-                x, y, z, laser_position[ilaser], properties, laserP[ilaser]
-            )
-            return Q * w  # Weighted source term
-
-        # Vectorize over all Level 3 elements
-        vstepcomputeCoarseSource = jax.vmap(stepcomputeCoarseSource)
-        _data = vstepcomputeCoarseSource(jnp.arange(ne_nn[1]))
-
-        # Integrate over elements using shape functions
-        _data1tmp = multiply(LevelShape[0], _data).sum(axis=1)
-        return _data1tmp
-
-    # Vectorize over all laser positions
-    vstepLaserPosition = jax.vmap(stepLaserPosition)
-
-    # Average source over all laser positions
-    lshape = laser_position.shape[0]
-    _data1 = vstepLaserPosition(jnp.arange(lshape)).sum(axis=0) / lshape
-
-    # Project integrated source to global vector of target level
-    return LevelShape[2] @ _data1.reshape(-1)
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
 def computeL1TprimeTerms_Part1(Levels, ne_nn, L3k, Shapes, L2k):
     """
     Compute the subgrid correction terms for Level 1 from Levels 2 and 3.
@@ -2958,61 +2163,6 @@ def computeL2Temperature(
 
 
 @partial(jax.jit, static_argnames=["ne_nn"])
-def computeSourcesL3(Level, v, ne_nn, properties, laserP):
-    """
-    Compute the integrated volumetric heat source for Level 3.
-
-    This function evaluates the laser heat source at quadrature points
-    using the Level 3 mesh and assembles the global source vector.
-
-    Parameters:
-    Level (dict): Level 3 mesh and field data.
-        - Level["elements"]: Element dimensions (nx, ny, nz).
-        - Level["nodes"]: Node dimensions (nx, ny, nz).
-    v (array): Current laser position.
-    ne_nn (tuple): Mesh metadata.
-        - ne_nn[1]: Number of elements in Level 3.
-        - ne_nn[4]: Total number of nodes in Level 3.
-    properties (dict): Material and process properties.
-    laserP (float): Laser power at the current position.
-
-    Returns:
-    array: Assembled global source vector for Level 3.
-    """
-    # Get shape functions and quadrature weights
-    coords = getSampleCoords(Level)
-    Nf, _, wqf = computeQuad3dFemShapeFunctions_jax(coords)
-
-    def stepcomputeCoarseSource(ieltf):
-        # Convert element index to 3D indices and global node indices
-        ix, iy, iz, idx = convert2XYZ(
-            ieltf,
-            Level["elements"][0],
-            Level["elements"][1],
-            Level["nodes"][0],
-            Level["nodes"][1],
-        )
-        # Get quadrature point coordinates for this element
-        x, y, z = getQuadratureCoords(Level, ix, iy, iz, Nf)
-        w = wqf
-
-        # Evaluate laser source at quadrature points
-        Q = computeSourceFunction_jax(x, y, z, v, properties, laserP)
-
-        # Integrate source over element using shape functions
-        return Nf @ Q * w, idx
-
-    # Vectorize over all Level 3 elements
-    vstepcomputeCoarseSource = jax.vmap(stepcomputeCoarseSource)
-    _data3, nodes3 = vstepcomputeCoarseSource(jnp.arange(ne_nn[1]))
-
-    # Assemble global source vector using node indices
-    Ff = bincount(nodes3.reshape(-1), _data3.reshape(-1), ne_nn[4])
-
-    return Ff
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
 def computeSolutions_L3(
     FinalL2, L2L3Interp, Levels, ne_nn, L3T0, L3F, L3k, L3rhocp, dt
 ):
@@ -3148,24 +2298,6 @@ def getSampleCoords(Level):
     return jnp.concatenate([x, y, z], axis=1)
 
 
-def getQuadratureCoords(Level, ix, iy, iz, Nf):
-    """
-    Computes the quadrature point coordinates for a given element.
-
-    Parameters:
-    Level (dict): Mesh data for a given level.
-    ix, iy, iz (int): Element indices in x, y, z directions.
-    Nf (array): Shape function matrix.
-
-    Returns:
-    tuple: x, y, z coordinates at quadrature points.
-    """
-    coords_x = Level["node_coords"][0][Level["connect"][0][ix, :]].reshape(-1, 1)
-    coords_y = Level["node_coords"][1][Level["connect"][1][iy, :]].reshape(-1, 1)
-    coords_z = Level["node_coords"][2][Level["connect"][2][iz, :]].reshape(-1, 1)
-    return Nf @ coords_x, Nf @ coords_y, Nf @ coords_z
-
-
 @partial(jax.jit, static_argnames=["ne_nn"])
 def computeL2TprimeTerms_Part2(Levels, ne_nn, L3Tp, L3Tp0, L3rhocp, dt, Shapes, L2V):
     """
@@ -3219,417 +2351,6 @@ def computeL2TprimeTerms_Part2(Levels, ne_nn, L3Tp, L3Tp0, L3rhocp, dt, Shapes, 
     L2V += Shapes[2][2] @ _data2.reshape(-1)
 
     return L2V
-
-
-@partial(jax.jit, static_argnames=["ne_nn", "tmp_ne_nn", "substrate", "subcycle"])
-def subcycleGOMELT(
-    Levels,
-    ne_nn,
-    Shapes,
-    substrate,
-    LInterp,
-    tmp_ne_nn,
-    laser_position,
-    properties,
-    laserP,
-    subcycle,
-    max_accum_L3,
-    accum_L3,
-):
-    """
-    Perform a full predictor-corrector subcycling step for the GO-MELT model.
-
-    This function executes the multilevel thermal update across Levels 1, 2, and 3,
-    using nested subcycling and subgrid scale corrections. It includes both the
-    predictor and corrector phases, updating temperature fields and phase states.
-
-    Parameters:
-    Levels (list of dict): Multilevel simulation state.
-    ne_nn (tuple): Mesh metadata for all levels.
-    Shapes (tuple): Shape function data for subgrid projections.
-    substrate (tuple): Substrate node indices for each level.
-    LInterp (tuple): Interpolation matrices between levels.
-    tmp_ne_nn (tuple): Level 1 metadata (element count, inactive node index).
-    laser_position (array): Laser positions for each substep.
-    properties (dict): Material and process properties.
-    laserP (array): Laser power at each substep.
-    subcycle (tuple): Subcycling parameters (L2 steps, L3 steps per L2, etc.).
-    max_accum_L3 (array): Max accumulated melt time for Level 3 nodes.
-    accum_L3 (array): Current accumulated melt time for Level 3 nodes.
-
-    Returns:
-    tuple: Updated Levels, Level 2 and 3 temperature histories, Level 3 Tprime history,
-           updated max_accum_L3 and accum_L3.
-    """
-    # --- Level 1 Material Properties ---
-    _, _, L3k_L1, L3rhocp_L1 = computeStateProperties(
-        Levels[3]["T0"], Levels[3]["S1"], properties, substrate[3]
-    )
-    _, _, L2k_L1, L2rhocp_L1 = computeStateProperties(
-        Levels[2]["T0"], Levels[2]["S1"], properties, substrate[2]
-    )
-
-    # Update Level 1 S1 from Level 2 overlap
-    _val = interpolatePoints(Levels[2], Levels[2]["S1"], Levels[2]["overlapCoords"])
-    _idx = getOverlapRegion(
-        Levels[2]["overlapNodes"], Levels[1]["nodes"][0], Levels[1]["nodes"][1]
-    )
-    Levels[1]["S1"] = Levels[1]["S1"].at[_idx].set(_val)
-    Levels[1]["S1"] = Levels[1]["S1"].at[: substrate[1]].set(1)
-
-    _, _, L1k, L1rhocp = computeStateProperties(
-        Levels[1]["T0"], Levels[1]["S1"], properties, substrate[1]
-    )
-
-    # --- Level 1 Source and Subgrid Terms ---
-    L1F = computeLevelSource(
-        Levels, ne_nn, laser_position, Shapes[1], properties, laserP
-    )
-    L1F = computeConvRadBC(
-        Levels[1], Levels[1]["T0"], tmp_ne_nn[0], ne_nn[2], properties, L1F
-    )
-    L1V = computeL1TprimeTerms_Part1(Levels, ne_nn, L3k_L1, Shapes, L2k_L1)
-
-    # --- Level 1 Temperature Predictor ---
-    L1T = computeL1Temperature(
-        Levels,
-        ne_nn,
-        tmp_ne_nn,
-        L1F,
-        L1V,
-        L1k,
-        L1rhocp,
-        laser_position[:, 5].sum(),
-        properties,
-    )
-    L1T = jnp.maximum(properties["T_amb"], L1T)  # TFSP
-
-    # --- Subcycle Level 2 ---
-    def subcycleL2_Part1(_L2carry, _L2sub):
-        # Compute interpolation weights for Level 2 boundary conditions
-        alpha_L2 = (_L2sub + 1) / subcycle[3]
-        beta_L2 = 1 - alpha_L2
-
-        # Determine the laser substeps for this Level 2 subcycle
-        Lidx = _L2sub * subcycle[1] + jnp.arange(subcycle[1])
-
-        # --- Material Properties ---
-        # Compute Level 3 properties using current Level 3 temperature and phase
-        _, _, L3k_L2, _ = computeStateProperties(
-            _L2carry[2], _L2carry[4], properties, substrate[3]
-        )
-
-        # Compute Level 2 properties using current Level 2 temperature and phase
-        L2S1, _, L2k, L2rhocp = computeStateProperties(
-            _L2carry[0], _L2carry[1], properties, substrate[2]
-        )
-
-        # --- Source Term ---
-        # Compute laser source term for Level 2 using Level 3 mesh
-        L2F = computeLevelSource(
-            Levels, ne_nn, laser_position[Lidx, :], Shapes[2], properties, laserP[Lidx]
-        )
-
-        # Add convection, radiation, and evaporation boundary conditions
-        L2F = computeConvRadBC(
-            Levels[2],
-            _L2carry[0],
-            ne_nn[0],
-            ne_nn[3],
-            properties,
-            L2F,
-        )
-
-        # --- Subgrid Correction ---
-        # Compute divergence of subgrid heat flux from Level 3 to Level 2
-        L2V = computeL2TprimeTerms_Part1(Levels, ne_nn, _L2carry[3], L3k_L2, Shapes)
-
-        # --- Temperature Solve ---
-        # Interpolate Level 1 temperature to Level 2 boundary using alpha-beta blend
-        _BC = alpha_L2 * L1T + beta_L2 * Levels[1]["T0"]
-
-        # Solve Level 2 temperature using matrix-free FEM
-        L2T = computeL2Temperature(
-            _BC,
-            LInterp[0],
-            Levels,
-            ne_nn,
-            _L2carry[0],
-            L2F,
-            L2V,
-            L2k,
-            L2rhocp,
-            laser_position[Lidx, 5].sum(),
-        )
-        L2T = jnp.maximum(properties["T_amb"], L2T)  # TFSP
-
-        # --- Subcycle Level 3 ---
-        def subcycleL3_Part1(_L3carry, _L3sub):
-            # Compute Level 3 material properties
-            L3S1, _, L3k, L3rhocp = computeStateProperties(
-                _L3carry[0], _L3carry[1], properties, substrate[3]
-            )
-
-            # Determine laser index for this Level 3 substep
-            LLidx = _L3sub + _L2sub * subcycle[1]
-
-            # Compute laser source term for Level 3
-            L3F = computeSourcesL3(
-                Levels[3], laser_position[LLidx, :], ne_nn, properties, laserP[LLidx]
-            )
-
-            # Add convection, radiation, and evaporation boundary conditions
-            L3F = computeConvRadBC(
-                Levels[3], _L3carry[0], ne_nn[1], ne_nn[4], properties, L3F
-            )
-
-            # Interpolate Level 2 temperature to Level 3 boundary
-            alpha_L3 = (_L3sub + 1) / subcycle[4]
-            beta_L3 = 1 - alpha_L3
-            _BC = alpha_L3 * L2T + beta_L3 * _L2carry[0]
-
-            # Solve Level 3 temperature
-            L3T = computeSolutions_L3(
-                _BC,
-                LInterp[1],
-                Levels,
-                ne_nn,
-                _L3carry[0],
-                L3F,
-                L3k,
-                L3rhocp,
-                laser_position[LLidx, 5],
-            )
-            L3T = jnp.maximum(properties["T_amb"], L3T)  # TFSP
-
-            return ([L3T, L3S1], [L3T, L3S1])
-
-        # Run Level 3 subcycling loop
-        [L3T, L3S1], _ = jax.lax.scan(
-            subcycleL3_Part1,
-            [_L2carry[2], _L2carry[4]],
-            jnp.arange(subcycle[1]),
-        )
-
-        # Compute Updated Level 3 Tprime and update Level 2 Temperature
-        L3Tp, L2T = getNewTprime(Levels[3], L3T, L2T, Levels[2], LInterp[1])
-
-        return ([L2T, L2S1, L3T, L3Tp, L3S1], [L2T, L2S1, L3T, L3Tp, L3S1])
-
-    # Run Level 2 subcycling loop
-    [L2T, _, _, L3Tp, _], [_, _, _, L3Tp_L2, _] = jax.lax.scan(
-        subcycleL2_Part1,
-        [
-            Levels[2]["T0"],
-            Levels[2]["S1"],
-            Levels[3]["T0"],
-            Levels[3]["Tprime0"],
-            Levels[3]["S1"],
-        ],
-        jnp.arange(subcycle[0]),
-    )
-
-    # --- Level 2 Tprime and Level 1 Corrector Update ---
-    L2Tp, L1T = getNewTprime(Levels[2], L2T, L1T, Levels[1], LInterp[0])
-    L1V = computeL1TprimeTerms_Part2(
-        Levels,
-        ne_nn,
-        L3Tp,
-        L2Tp,
-        L3rhocp_L1,
-        L2rhocp_L1,
-        laser_position[:, 5].sum(),
-        Shapes,
-        L1V,
-    )
-    L1T = computeL1Temperature(
-        Levels,
-        ne_nn,
-        tmp_ne_nn,
-        L1F,
-        L1V,
-        L1k,
-        L1rhocp,
-        laser_position[:, 5].sum(),
-        properties,
-    )
-    L1T = jnp.maximum(properties["T_amb"], L1T)  # TFSP
-
-    # --- Subcycle Level 2 ---
-    def subcycleL2_Part2(_L2carry, _L2sub):
-        # Compute interpolation weights for Level 2 boundary conditions
-        alpha_L2 = (_L2sub + 1) / subcycle[3]
-        beta_L2 = 1 - alpha_L2
-
-        # Determine the laser substeps for this Level 2 subcycle
-        Lidx = _L2sub * subcycle[1] + jnp.arange(subcycle[1])
-
-        # --- Material Properties ---
-        # Compute Level 3 properties using current Level 3 temperature and phase
-        _, _, L3k_L2, L3rhocp_L2 = computeStateProperties(
-            _L2carry[2], _L2carry[4], properties, substrate[3]
-        )
-
-        # Compute Level 2 properties using current Level 2 temperature and phase
-        L2S1, _, L2k, L2rhocp = computeStateProperties(
-            _L2carry[0], _L2carry[1], properties, substrate[2]
-        )
-
-        # --- Source Term ---
-        # Compute laser source term for Level 2 using Level 3 mesh
-        L2F = computeLevelSource(
-            Levels, ne_nn, laser_position[Lidx, :], Shapes[2], properties, laserP[Lidx]
-        )
-
-        # Add convection, radiation, and evaporation boundary conditions
-        L2F = computeConvRadBC(
-            Levels[2],
-            _L2carry[0],
-            ne_nn[0],
-            ne_nn[3],
-            properties,
-            L2F,
-        )
-
-        # --- Subgrid Correction ---
-        # Compute divergence of subgrid heat flux from Level 3 to Level 2
-        L2V = computeL2TprimeTerms_Part1(Levels, ne_nn, _L2carry[3], L3k_L2, Shapes)
-
-        # Add time derivative correction from Level 3 to Level 2
-        L2V = computeL2TprimeTerms_Part2(
-            Levels,
-            ne_nn,
-            L3Tp_L2[_L2sub],
-            _L2carry[3],
-            L3rhocp_L2,
-            laser_position[Lidx, 5].sum(),
-            Shapes,
-            L2V,
-        )
-
-        # --- Temperature Solve ---
-        # Interpolate Level 1 temperature to Level 2 boundary using alpha-beta blend
-        _BC = alpha_L2 * L1T + beta_L2 * Levels[1]["T0"]
-
-        # Solve Level 2 temperature using matrix-free FEM
-        L2T = computeL2Temperature(
-            _BC,
-            LInterp[0],
-            Levels,
-            ne_nn,
-            _L2carry[0],
-            L2F,
-            L2V,
-            L2k,
-            L2rhocp,
-            laser_position[Lidx, 5].sum(),
-        )
-        L2T = jnp.maximum(properties["T_amb"], L2T)  # TFSP
-
-        # --- Subcycle Level 3 ---
-        def subcycleL3_Part2(_L3carry, _L3sub):
-            # Compute Level 3 material properties
-            L3S1, L3S2, L3k, L3rhocp = computeStateProperties(
-                _L3carry[0], _L3carry[1], properties, substrate[3]
-            )
-
-            # Determine laser index for this Level 3 substep
-            LLidx = _L3sub + _L2sub * subcycle[1]
-
-            # Compute laser source term for Level 3
-            L3F = computeSourcesL3(
-                Levels[3], laser_position[LLidx, :], ne_nn, properties, laserP[LLidx]
-            )
-
-            # Add convection, radiation, and evaporation boundary conditions
-            L3F = computeConvRadBC(
-                Levels[3], _L3carry[0], ne_nn[1], ne_nn[4], properties, L3F
-            )
-
-            # Interpolate Level 2 temperature to Level 3 boundary
-            alpha_L3 = (_L3sub + 1) / subcycle[4]
-            beta_L3 = 1 - alpha_L3
-            _BC = alpha_L3 * L2T + beta_L3 * _L2carry[0]
-
-            # Solve Level 3 temperature
-            L3T = computeSolutions_L3(
-                _BC,
-                LInterp[1],
-                Levels,
-                ne_nn,
-                _L3carry[0],
-                L3F,
-                L3k,
-                L3rhocp,
-                laser_position[LLidx, 5],
-            )
-            L3T = jnp.maximum(properties["T_amb"], L3T)  # TFSP
-
-            # --- Accumulated Melt Time Update ---
-            # Reset accumulation if solid becomes liquid again
-            _resetmask = ((1 - 2 * _L3carry[2]) * L3S2) == 1
-            _resetaccumtime = _L3carry[4] * _resetmask
-
-            # Update max accumulated melt time
-            _max_check = jnp.maximum(_resetaccumtime, _L3carry[3])
-            max_accum_L3 = _max_check
-
-            # Update current accumulated melt time
-            accum_L3 = _L3carry[4] + laser_position[LLidx, 5] * L3S2 - _resetaccumtime
-
-            return (
-                [L3T, L3S1, L3S2, max_accum_L3, accum_L3],
-                [L3T, L3S1, L3S2, max_accum_L3, accum_L3],
-            )
-
-        # Run Level 3 subcycling loop
-        [L3T, L3S1, L3S2, max_accum_L3, accum_L3], _ = jax.lax.scan(
-            subcycleL3_Part2,
-            [_L2carry[2], _L2carry[4], _L2carry[5], _L2carry[6], _L2carry[7]],
-            jnp.arange(subcycle[1]),
-        )
-
-        # Compute updated Tprime for Level 3 and temperature for Level 2
-        L3Tp, L2T = getNewTprime(Levels[3], L3T, L2T, Levels[2], LInterp[1])
-
-        return (
-            [L2T, L2S1, L3T, L3Tp, L3S1, L3S2, max_accum_L3, accum_L3],
-            [L2T, L2S1, L3T, L3Tp, L3S1, L3S2, max_accum_L3, accum_L3],
-        )
-
-    [
-        Levels[2]["T0"],
-        Levels[2]["S1"],
-        Levels[3]["T0"],
-        Levels[3]["Tprime0"],
-        Levels[3]["S1"],
-        Levels[3]["S2"],
-        max_accum_L3,
-        accum_L3,
-    ], [L2all, _, L3all, L3pall, _, _, _, _] = jax.lax.scan(
-        subcycleL2_Part2,
-        [
-            Levels[2]["T0"],
-            Levels[2]["S1"],
-            Levels[3]["T0"],
-            Levels[3]["Tprime0"],
-            Levels[3]["S1"],
-            Levels[3]["S2"],
-            max_accum_L3,
-            accum_L3,
-        ],
-        jnp.arange(subcycle[0]),
-    )
-
-    # --- Final Tprime and Phase Updates ---
-    Levels[2]["Tprime0"], Levels[1]["T0"] = getNewTprime(
-        Levels[2], Levels[2]["T0"], L1T, Levels[1], LInterp[0]
-    )
-    Levels[0]["S1"] = Levels[0]["S1"].at[Levels[0]["idx"]].set(Levels[3]["S1"])
-    Levels[0]["S2"] = Levels[0]["S2"].at[:].set(False)
-    Levels[0]["S2"] = Levels[0]["S2"].at[Levels[0]["idx"]].set(Levels[3]["S2"])
-
-    return Levels, L2all, L3all, L3pall, max_accum_L3, accum_L3
 
 
 def printLevelMaxMin(Ls, Lnames):
